==> 03_map.ipynb <==
#pip install -U folium

import pandas as pd
df = pd.read_csv(r"C:\Users\Gabor\Documents\03_Programming\04_Projects\01_WCS-cities\\WCS-members.csv", sep = ";", decimal = ",")
df = df.drop([3], axis=0) # drop Vienna
df = df[df["tagok"] > 0]
df.head()

import folium
# https://python-visualization.github.io/folium/modules.html
 
SF_COORDINATES = (52.3112, 13.241800)
 
# create empty map zoomed in on San Francisco
map = folium.Map(location=SF_COORDINATES, zoom_start=4.3)

# I can add marker one by one on the map
for i in range(0,len(df)):
   folium.Circle(
      location=[df.iloc[i]['latitude'], df.iloc[i]['longitude']],
      popup = df.iloc[i]['varos'],
      tooltip = df.iloc[i]['varos'] + ": " + str(df.iloc[i]['tagok']),
      #tooltip = str(df.iloc[i]['tagok']),
      radius=float(df.iloc[i]['tagok']*50),
      color='crimson',
      fill=True,
      fill_color='crimson'
   ).add_to(map)

map

# kezeljuk a missing tagszamot, aztan johet a wiki apizasa.
